Update patch set 1
Patch Set 1: Code-Review-1

As others have said, it'd be best to keep this looking like it does on Nexus devices. Plus, if it /was/ due to not having RAW, then it would make sense to not put this icon back unless RAW is actually added. If RAW was added, then I personally would be fine with this icon being used instead of the one being used now.

Patch-set: 1
Label: Code-Review=-1