Update patch set 2
Patch Set 2: Verified+1 Code-Review+2

FYI if you've enabled this and suddenly your preview freezes after a shot, it means that you're not actually doing ZSL and the camera driver is stopping your preview.

I noticed that the Google Edition of the M7 and JF do this. There may be additional incantations needed. On JF, the magic raw command used to enabled the correct behavior but it does not on the GED driver.

This works as advertised on HLTE and others though.

Patch-set: 2
Label: Verified=+1