Update patch set 4
Patch Set 4:

hey Michael

lol that was not me who answered you on github lol

To answer your questions:

Why I use not as privileged app anymore is simple...at least for the nexus devices for user who flash PhotoSphere camera the available gapps get confused due that they do not find the correct path anymore

sign it with the platform cert is doing the same and you are able to overwrite the Power button and keep the app in system/app

Your question to fallback system...at least on slim a lot user like to use the camera app on another roms...in this case it will crash of course due of security exception. Added simply a try catch in Camera activity to that power button feature is disabled in this case.

Not really needed but I like to catch all cases :p

Patch-set: 4