Update patch set 1
Patch Set 1: Code-Review+1

I failed to see this before submitting a similar patch
http://review.cyanogenmod.org/#/c/59844/

Happens for White Balance too on the unoffical SM-T310 build and presumably the kona too which has tungsten white-balance disabled.

The additional problem my patch addresses is with the onScreenIndicator which is also out of sync with the filtered preferences.

Patch-set: 1