Update patch set 1
Patch Set 1:

A day should be fine, but its not necessary really. The reproduction scenario is to cause a mismatch of keys. That is, if you Camera has platform signature currently, and without this patch, you pushed Camera without the platform signature, you'd see the exception. 

By changing the protection level the platform/non-platform signature portion no longer plays a factor, and brings everyone back to a normal behavior.

Patch-set: 1