Update patch set 2
Patch Set 2:

There was some discussion on this lately.

Personally, I hate it.

But as an advanced feature, on a camera with lots of options, you might want to see the real composition instead of cropping to fit the aspect ratio.

I think we need to be more thoughtful about it though and where we are lumping all these advanced features.

Patch-set: 2
Reviewer: Steve Kondik <32@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0