Update patch set 2
Patch Set 2:

The transition from the viewfinder to the black background is jarring and looks broken, basically. There's nothing to stop us fixing this cosmetic issue, an attractive border and background would help a lot, as well as the possibility of using that real estate for UI that doesn't interfere with the viewfinder. 

If only manufacturers would match their camera aspect ratios to the device display :D

Patch-set: 2
Label: Code-Review=0