Update patch set 1
Patch Set 1: Code-Review+1

Just a question about 1): is it possible to present to the MTP interface a directory with a different name?

i.e. make it so that within the phone you see "Camera" (which makes a lot more sense), but from MTP you see "100ANDRO".

This change makes browsing for photos a bit less immediate.

Patch-set: 1