Update patch set 4
Patch Set 4:

When we want to follow all the specifications the full path to an image should be like "DCIM/100ANDRO/IMAG{0001 - 9999}.JPG".

But when we want to follow these rules we also have to handle the case that a user takes more than 9999 images. In that case a new directory has to be created in the "DCIM" folder ("101ANDRO" for example) and the images names in this directory have to start with "0001" again. And so on and on... ;)

Patch-set: 4
Reviewer: Jacob Müller <4884@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0