Update patch set 2
Patch Set 2:

Hmm, not sure what to think of it. Yes, the shutter isn't (partly) below the navigation bar with this patch, but the UI controls feel pretty 'large' now that they aren't really attached to the edge anymore.

We definitely should fix the margins of the menu controls though, to correctly center align the buttons.

Patch-set: 2