Update patch set 5
Patch Set 5:

I'm fine with this change if accompanied with the alignment fix. The upper circle is misaligned in the "After" screenshot.

Was there any reason for the current size and placement? I've not used a device without software buttons in a while and when I did I didn't use the camera that much.

Patch-set: 5