Update patch set 2
Patch Set 2: Code-Review+1

From eglDestroyContext spec: "If the EGL rendering context context is not current to any thread, eglDestroyContext destroys it immediately. Otherwise, context is destroyed when it becomes not current to any thread."

eglDestroySurface: "If the EGL surface surface is not current to any thread, eglDestroySurface destroys it immediately. Otherwise, surface is destroyed when it becomes not current to any thread. Furthermore, resources associated with a pbuffer surface are not released until all color buffers of that pbuffer bound to a texture object have been released."

Broadcom's VideoCoreIV doesn't seem to implement the specifications properly - it treats current vs non-current contexts the same - so it's better to reorder the EGL calls in the optimal order. May also help on other devices that rely on prebuilt EGL drivers.

Patch-set: 2
Label: Code-Review=+1