Update patch set 5
Patch Set 5:

Unless there is a significant performance improvement with this, I would say that reverting is a good choice.

The other choice of having the openCamera/preview thread waiting for the UI initialization to complete would completely destroy the purpose of this patch.

Moving mFocusManager.setPreviewSize back to onCameraOpened might fix this, but there has to be a reason why it was moved from there in the first place.

Patch-set: 5