Update patch set 2
Patch Set 2: Code-Review-1

Arne, I didn't test with this reverted because I agree with Steve that this shouldn't be reverted (because most likely we are dealing with buggy vendor code).  I added comments here because it seemed like a reasonable place to get feedback and it is likely the same underlying problem.

I definitely did not mean to imply I wanted it reverted.

Patch-set: 2
Label: Code-Review=-1