Update patch set 2
Patch Set 2:

I spent a little time trying to see if I could make our HAL not broken and didn't have much luck.  Instead, I'm suggesting conditionally disabling the threaded startup.

I added a resource to allow devices to opt-out of the threaded startup:

http://review.cyanogenmod.org/#/c/76690/
http://review.cyanogenmod.org/#/c/76691/

which fixes the camera on picassowifi and doesn't break it on a qcom device that I tested.

FYI, this is actually a blocker for M12 for picassowifi (and i9300? i9500?) so anything we can do to get it resolved would be greatly appreciated.

Patch-set: 2