Update patch set 1
Patch Set 1:

Yes. Without this the camera on d2 pauses to do a full focus cycle and the flash actually fires *after* the shot. Camera2 uses auto mode for TTF and continuous focus if a device supports it otherwise. There is no reason to do a focus when continuous focus is running.

Patch-set: 1