Update patch set 2
Patch Set 2:

I don't think so... Because, then the last item can't use the undo option. I think the logic should be applied to all the items, and right now it leaving 1 item always outside. It should be if there is 1 or more pictures (or if there are more than 0 pictures) you can use the undo pictures. Right now the logic is if there are 2 or more pictures, which doesn't make sense.

Patch-set: 2