Update patch set 1
Patch Set 1:

It crashes on a Galaxy S2 i9100.
I only renamed the files and the reference in the xml-file. This is because the name of the xml-file somehow in this situation creates a circular reference with the png-files. I don't know why specific for this device. In LogCat it reports an invalid src-attribute which made also no sense, but works after simply renaming the files and reference.

Patch-set: 1
Reviewer: Dominggoes Isakh <13873@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0