Camera2: override capture support level
-The capture support level detected for our Camera is
LIMITED_YUV which leads to incorrect picture sizes. Well,
they might be for YUV but our Cameras are capable of
LIMITED_JPEG which corrects the camera issues. GservicesHelper
seems to be a place where this detection can be overriden using
a ContentResolver. I thought it would be simpler to just set
a property.

-CaptureSupportLevel is an enum with these values:
ZSL = 1
LEGACY_JPEG = 2
LIMITED_JPEG = 3
LIMITED_YUV = 4
and you can use -1 for unused. The property should correspond
to one of these values.

Change-Id: I72ad06572ecc6fbc7b016f53d2e8a0f50cbe3b6a