Fix incorrect use of Application Context for display info
An Application object should not be used to query display information,
such as orientation or dimensions, since an Application may have
multiple activities, and each Activity may be on a different display
or multi-window situation.

The platform strict mode now enforces this by disallowing fetching a
WindowManager instance using an Application context, leading to app
crashes in some testing scenarios.

Update all queries that retrieve a WindowManager to use an Activity,
and plumb that through all the necessary layers.

Bug: 163823707
Test: Camera2 app doesn't crash on test device on master ToT when switching
    to video mode

Change-Id: Id97c992e1ff99fe161b786166008011228856030