Camera: Remove possible duplicate filmstrip items
Due to timing the filmstrip items in the data adapter
can be refreshed and include an item which still has
a placeholder that can receive an update. This can
cause two items to point to the same media
location. Check for an older filmstrip item with
the same URI and remove it if necessary.
Additionally, fix a few strict mode violations.

Bug: 195337468
Test: Manual using camera application
Change-Id: I953dc38f460718d8fb76c48910bbae07f0592e45