Camera2 (AOSP Camera app): Support split-screen in picture mode.
Now that the onCameraAccessPrioritiesChanged callback is correctly
triggering, the Camera2 app should demonstrate this capability. Some
minor changes have been implemented to support this. The app now
anticipates getting disconnected from the camera while in the foreground
and will try to obtain a connection when not paused and
onCameraAccessPrioritiesChanged is called.

Bug: 224887921
Test: Camera access is consistently transferred when tapping (x20)
Test: Camera access is consistently transferred when switching the camera (x20)
Test: Switching between video and picture mode works outside of split screen.
Test: Pausing and resuming the app does not result in a crash (x20)
Test: Rapidly switched focus between two camera apps, verified that does not cause issues
Change-Id: I334a79f112da31f9935b81797543e862bf1a6e43