Add required flag to receiver in VideoModule
Android T allows apps to declare a runtime receiver as not exported
by invoking registerReceiver with a new RECEIVER_NOT_EXPORTED flag;
receivers registered with this flag will only receive broadcasts from
the platform and the app itself. However to ensure developers can
properly protect their receivers, all apps targeting U
and registering a receiver for non-system broadcasts must
specify either the exported or not exported flag when invoking
registerReceiver; if one of these flags is not provided, the
platform will throw a SecurityException. This commit updates
the receiver with the RECEIVER_EXPORTED flag since the
broadcast is sent from the media provider.

Bug: 234659204
Test: Build
Change-Id: I227e8d9e2ace18119a37ddb0c2c4ddf4c3be6209